BatBot initial testing and integration (#305)
* initial testing and integration

* batbot metadata parser

* batbot spectrogram generation

* remove old spectrogram generation code

* swap back to using the github installation for batbot

* use temp branch for start/stop fixes

* increase accuracy for spectrograms and annotations

* thumbnail centering fixes

* add noise filter

* contour testing

* update batbot

* fix NABat spectrogram generation

* client linting

* removing integration notes

* Update client/src/components/TransparencyFilterControl.vue


* reverting float to ints for pixel fields

* remove uneeded dependencies

* import GRTS updated for sciencebase.gov downtime

* add batbot issue for ml model integration

* main merge migration update

* swap to port 8080 for client based on main's client redirect port

* remove matplotlib from importRecordings

* Update redirect URL for local developers to 8080

* adjust default zoom size

---------

+++ bats_ai/core/models/nabat/nabat_spectrogram.py
@@ -20,9 +20,9 @@ class NABatSpectrogram(TimeStampedModel, models.Model):
     images = GenericRelation(SpectrogramImage)
     width = models.IntegerField()  # pixels
     height = models.IntegerField()  # pixels
-    duration = models.IntegerField()  # milliseconds
-    frequency_min = models.IntegerField()  # hz
-    frequency_max = models.IntegerField()  # hz
+    duration = models.FloatField()  # milliseconds
+    frequency_min = models.FloatField()  # hz
+    frequency_max = models.FloatField()  # hz
 
     @property
     def image_url_list(self):
